+++ src/components/cell-control/autocompleter.jsx
@@ -2,12 +2,12 @@ import PropTypes from 'prop-types';
 import React, { forwardRef, useRef } from 'react';
 import CellControl from './cell-control.jsx';
 import AutocompleterControl from '../control/autocompleter.jsx';
-import styles from './select.css';
+import styles from './autocompleter.css';
 
 const Autocompleter = forwardRef(function Autocompleter(props, ref) {
   const classNames = {
-    container: undefined,
-    innerContainer: styles.container,
+    container: styles.container,
+    innerContainer: styles.innerContainer,
     input: styles.input,
     invalidInput: styles.invalidInput,
     ...props.classNames,
@@ -27,8 +27,8 @@ const Autocompleter = forwardRef(function Autocompleter(props, ref) {
         apiEndpoint={props.apiEndpoint}
         classNames={{
           container: classNames.innerContainer,
-          input: styles.input,
-          invalidInput: styles.invalidInput,
+          input: classNames.input,
+          invalidInput: classNames.invalidInput,
         }}
         displayValueEvaluator={props.displayValueEvaluator}
         id={props.id}